test: improve cast mocks naming
refs: https://github.com/RonasIT/laravel-helpers/issues/263
