+++ tests/fixtures/ModelTestStateTest/changes_equals_fixture_with_castable/changed_dataset.json
@@ -2,36 +2,28 @@
   {
     "id": 1,
     "name": "name1",
-    "castable_field": 1,
-    "address_line_one": "100 Main St",
-    "address_line_two": "Apt 1",
+    "address": "{\"line_one\": \"100 Main St\", \"line_two\": \"Apt 1\"}",
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   },
   {
     "id": 2,
     "name": "name2",
-    "castable_field": 2,
-    "address_line_one": "200 Oak Ave Updated",
-    "address_line_two": "Suite 20",
+    "address": "{\"line_one\": \"200 Oak Ave Updated\", \"line_two\": \"Suite 20\"}",
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   },
   {
     "id": 3,
     "name": "name3_updated",
-    "castable_field": 2,
-    "address_line_one": "300 Pine Rd Updated",
-    "address_line_two": "Floor 3",
+    "address": "{\"line_one\": \"300 Pine Rd Updated\", \"line_two\": \"Floor 3\"}",
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   },
   {
     "id": 5,
     "name": "name5",
-    "castable_field": 1,
-    "address_line_one": "500 New Blvd",
-    "address_line_two": "Lobby",
+    "address": "{\"line_one\": \"500 New Blvd\", \"line_two\": \"Lobby\"}",
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   }
