+++ tests/fixtures/ModelTestStateTest/changes_equals_fixture_with_castable/initial_dataset.json
@@ -2,36 +2,28 @@
   {
     "id": 1,
     "name": "name1",
-    "castable_field": 1,
-    "address_line_one": "100 Main St",
-    "address_line_two": "Apt 1",
+    "address": "{\"line_one\": \"100 Main St\", \"line_two\": \"Apt 1\"}",
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   },
   {
     "id": 2,
     "name": "name2",
-    "castable_field": 1,
-    "address_line_one": "200 Oak Ave",
-    "address_line_two": "Suite 10",
+    "address": "{\"line_one\": \"200 Oak Ave\", \"line_two\": \"Suite 10\"}",
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   },
   {
     "id": 3,
     "name": "name3",
-    "castable_field": 1,
-    "address_line_one": "300 Pine Rd",
-    "address_line_two": "Floor 2",
+    "address": "{\"line_one\": \"300 Pine Rd\", \"line_two\": \"Floor 2\"}",
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   },
   {
     "id": 4,
     "name": "name4",
-    "castable_field": 1,
-    "address_line_one": "400 Elm St",
-    "address_line_two": "Unit 5",
+    "address": "{\"line_one\": \"400 Elm St\", \"line_two\": \"Unit 5\"}",
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   }
