+++ tests/fixtures/ModelTestStateTest/changes_equals_fixture_with_custom_parameterized_cast/changed_dataset.json
@@ -2,28 +2,28 @@
   {
     "id": 1,
     "name": "name1",
-    "castable_field": 50000,
+    "amount": 50000,
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   },
   {
     "id": 2,
     "name": "name2",
-    "castable_field": 150000,
+    "amount": 150000,
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   },
   {
     "id": 3,
     "name": "name3_updated",
-    "castable_field": 250000,
+    "amount": 250000,
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   },
   {
     "id": 5,
     "name": "name5",
-    "castable_field": 75000,
+    "amount": 75000,
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   }
