+++ tests/fixtures/ModelTestStateTest/changes_equals_fixture_with_custom_parameterized_cast/initial_dataset.json
@@ -2,28 +2,28 @@
   {
     "id": 1,
     "name": "name1",
-    "castable_field": 50000,
+    "amount": 50000,
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   },
   {
     "id": 2,
     "name": "name2",
-    "castable_field": 100000,
+    "amount": 100000,
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   },
   {
     "id": 3,
     "name": "name3",
-    "castable_field": 200000,
+    "amount": 200000,
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   },
   {
     "id": 4,
     "name": "name4",
-    "castable_field": 300000,
+    "amount": 300000,
     "created_at": "2018-10-10 10:10:10",
     "updated_at": "2018-10-10 10:10:10"
   }
