+++ tests/fixtures/ModelTestStateTest/db_changes/test_model_with_custom_parameterized_casts/assertion_fixture.json
@@ -2,19 +2,19 @@
     "updated": [
         {
             "id": 2,
-            "castable_field": "1,500.00"
+            "amount": "1,500.00"
         },
         {
             "id": 3,
             "name": "name3_updated",
-            "castable_field": "2,500.00"
+            "amount": "2,500.00"
         }
     ],
     "created": [
         {
             "id": 5,
             "name": "name5",
-            "castable_field": "750.00",
+            "amount": "750.00",
             "created_at": "2018-10-10 10:10:10",
             "updated_at": "2018-10-10 10:10:10"
         }
@@ -23,7 +23,7 @@
         {
             "id": 4,
             "name": "name4",
-            "castable_field": "3,000.00",
+            "amount": "3,000.00",
             "created_at": "2018-10-10 10:10:10",
             "updated_at": "2018-10-10 10:10:10"
         }
