+++ tests/support/Mock/Casts/AddressCastable.php
@@ -7,7 +7,7 @@ use Illuminate\Contracts\Database\Eloquent\CastsAttributes;
 use Illuminate\Database\Eloquent\Model;
 use JsonSerializable;
 
-readonly class CustomCastable implements Castable, JsonSerializable
+readonly class AddressCastable implements Castable, JsonSerializable
 {
     public function __construct(
         public string $lineOne,
@@ -26,20 +26,22 @@ readonly class CustomCastable implements Castable, JsonSerializable
     public static function castUsing(array $arguments): CastsAttributes
     {
         return new class implements CastsAttributes {
-            public function get(Model $model, string $key, mixed $value, array $attributes): CustomCastable
+            public function get(Model $model, string $key, mixed $value, array $attributes): AddressCastable
             {
-                return new CustomCastable(
-                    lineOne: $attributes['address_line_one'],
-                    lineTwo: $attributes['address_line_two'],
+                $data = json_decode($value, true) ?? [];
+
+                return new AddressCastable(
+                    lineOne: $data['line_one'] ?? '',
+                    lineTwo: $data['line_two'] ?? '',
                 );
             }
 
-            public function set(Model $model, string $key, mixed $value, array $attributes): array
+            public function set(Model $model, string $key, mixed $value, array $attributes): string
             {
-                return [
-                    'address_line_one' => $value->lineOne,
-                    'address_line_two' => $value->lineTwo,
-                ];
+                return json_encode([
+                    'line_one' => $value->lineOne,
+                    'line_two' => $value->lineTwo,
+                ]);
             }
         };
     }
