+++ tests/support/Mock/Models/TestModelNonIdPrimaryKey.php
@@ -5,7 +5,7 @@ namespace RonasIT\Support\Tests\Support\Mock\Models;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use RonasIT\Support\Tests\Support\Mock\Casts\JSONCustomCast;
+use RonasIT\Support\Tests\Support\Mock\Casts\UserSettingCast;
 use RonasIT\Support\Traits\ModelTrait;
 
 class TestModelNonIdPrimaryKey extends Model
@@ -22,7 +22,7 @@ class TestModelNonIdPrimaryKey extends Model
 
     protected $casts = [
         'json_field' => 'array',
-        'castable_field' => JSONCustomCast::class,
+        'castable_field' => UserSettingCast::class,
     ];
 
     public function relation(): HasMany
