+++ tests/support/Mock/Models/TestModelWithCasts.php
@@ -3,7 +3,7 @@
 namespace RonasIT\Support\Tests\Support\Mock\Models;
 
 use Illuminate\Database\Eloquent\Model;
-use RonasIT\Support\Tests\Support\Mock\Casts\JSONCustomCast;
+use RonasIT\Support\Tests\Support\Mock\Casts\UserSettingCast;
 use RonasIT\Support\Traits\ModelTrait;
 
 class TestModelWithCasts extends Model
@@ -29,6 +29,6 @@ class TestModelWithCasts extends Model
         'collection_field' => 'collection',
 
         // Custom CastsAttributes implementation
-        'castable_field' => JSONCustomCast::class,
+        'castable_field' => UserSettingCast::class,
     ];
 }
