+++ tests/support/Mock/Models/TestModelWithCustomParameterizedCast.php
@@ -3,16 +3,16 @@
 namespace RonasIT\Support\Tests\Support\Mock\Models;
 
 use Illuminate\Database\Eloquent\Model;
-use RonasIT\Support\Tests\Support\Mock\Casts\JSONCustomParameterizedCast;
+use RonasIT\Support\Tests\Support\Mock\Casts\DecimalAmountCast;
 
 class TestModelWithCustomParameterizedCast extends Model
 {
     protected $fillable = [
         'name',
-        'castable_field',
+        'amount',
     ];
 
     protected $casts = [
-        'castable_field' => JSONCustomParameterizedCast::class . ':2',
+        'amount' => DecimalAmountCast::class . ':2',
     ];
 }
