+++ tests/support/Mock/Models/TestModelWithDifferentTimestampNames.php
@@ -4,7 +4,7 @@ namespace RonasIT\Support\Tests\Support\Mock\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
-use RonasIT\Support\Tests\Support\Mock\Casts\JSONCustomCast;
+use RonasIT\Support\Tests\Support\Mock\Casts\UserSettingCast;
 use RonasIT\Support\Traits\ModelTrait;
 
 class TestModelWithDifferentTimestampNames extends Model
@@ -28,6 +28,6 @@ class TestModelWithDifferentTimestampNames extends Model
 
     protected $casts = [
         'json_field' => 'array',
-        'castable_field' => JSONCustomCast::class,
+        'castable_field' => UserSettingCast::class,
     ];
 }
